got rid of priorities
- moved some stuff to manager
- added __c_ctor__ and __c_dtor__
