+++ source/base_timer.cpp
@@ -44,7 +44,7 @@ namespace internal
 {
 
 //============================================================================//
-init_priority(110)
+
 base_timer::mutex_map_t base_timer::f_mutex_map;
 
 //============================================================================//
